requirements.txt
fastapi==0.121.1
uvicorn[standard]==0.38.0
python-multipart==0.0.20
python-dotenv==1.2.1
firebase-admin==6.5.0
google-cloud-firestore==2.18.0
openai>=1.55.0
